refactor: move the AMM code from zoe to run-protocol
