+++ packages/run-protocol/exported.js
@@ -1,1 +1,2 @@
 import './src/vaultFactory/types.js';
+import './src/vpool-xyk-amm/types.js';
