+++ packages/run-protocol/src/exported.js
@@ -1,1 +1,2 @@
+import './vpool-xyk-amm/types.js';
 import './vaultFactory/types.js';
