+++ packages/zoe/src/contracts/exported.js
@@ -3,4 +3,3 @@ import './loan/types.js';
 import './priceAggregatorTypes.js';
 import './callSpread/types.js';
 import './attestation/types.js';
-import './vpool-xyk-amm/types.js';
